Fixing known bugs that cause issues with testing

+++ tests/test_NIRCam.py
@@ -13,6 +13,12 @@ from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
 from eureka.S6_planet_spectra import s6_spectra as s6
 
+try:
+    import harmonica  # noqa: F401
+    harmonica_installed = True
+except ModuleNotFoundError:
+    harmonica_installed = False
+
 
 def test_NIRCam(capsys):
     with capsys.disabled():
@@ -115,30 +121,31 @@ def test_NIRCam(capsys):
     assert os.path.exists(name+os.sep+'figs')
 
     # Rerun Stages 5 and 6 using Harmonica
-    # remove Stage 5 and 6 temporary files
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}Stage5")
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}Stage6")
+    if harmonica_installed:
+        # remove Stage 5 and 6 temporary files
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}Stage5")
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}Stage6")
 
-    ecf_path = f'.{os.sep}NIRCam_ecfs{os.sep}Harmonica{os.sep}'
-    s5_meta = s5.fitlc(meta.eventlabel, ecf_path=ecf_path, s4_meta=None)
-    s6_meta, s6_lc = s6.plot_spectra(meta.eventlabel, ecf_path=ecf_path,
-                                     s5_meta=s5_meta)
+        ecf_path = f'.{os.sep}NIRCam_ecfs{os.sep}Harmonica{os.sep}'
+        s5_meta = s5.fitlc(meta.eventlabel, ecf_path=ecf_path, s4_meta=None)
+        s6_meta, s6_lc = s6.plot_spectra(meta.eventlabel, ecf_path=ecf_path,
+                                         s5_meta=s5_meta)
 
-    # run assertions for S5
-    meta.outputdir_raw = (f'data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}'
-                          f'Stage5{os.sep}')
-    name = pathdirectory(meta, 'S5', 1, ap=8, bg=12,
-                         old_datetime=s5_meta.datetime)
-    assert os.path.exists(name)
-    assert os.path.exists(name+os.sep+'figs')
+        # run assertions for S5
+        meta.outputdir_raw = (f'data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}'
+                              f'Stage5{os.sep}')
+        name = pathdirectory(meta, 'S5', 1, ap=8, bg=12,
+                             old_datetime=s5_meta.datetime)
+        assert os.path.exists(name)
+        assert os.path.exists(name+os.sep+'figs')
 
-    # run assertions for S6
-    meta.outputdir_raw = (f'data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}'
-                          f'Stage6{os.sep}')
-    name = pathdirectory(meta, 'S6', 1, ap=8, bg=12,
-                         old_datetime=s6_meta.datetime)
-    assert os.path.exists(name)
-    assert os.path.exists(name+os.sep+'figs')
+        # run assertions for S6
+        meta.outputdir_raw = (f'data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}'
+                              f'Stage6{os.sep}')
+        name = pathdirectory(meta, 'S6', 1, ap=8, bg=12,
+                             old_datetime=s6_meta.datetime)
+        assert os.path.exists(name)
+        assert os.path.exists(name+os.sep+'figs')
 
     # remove temporary files
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRCam{os.sep}Stage3")
